+++ fluidSimLib/gridSystemData3.h
@@ -35,9 +35,9 @@ public:
 
     boundingBox3 boundingBox() const;
 
-    size_t addScalarGrid3Data(const scalarGrid3Ptr scalarGrid3Data, double initValue = 0.0);
+    size_t addScalarGrid3Data(const scalarGrid3Ptr scalarGrid3Data, FloatType initValue = 0.0);
     size_t addVectorGrid3Data(const vectorGrid3Ptr vectorGrid3Data, const vector3& initValue = vector3());
-    size_t addAdvectableScalarData(const scalarGrid3Ptr& scalarGrid3Data, double initialVal = 0.0);
+    size_t addAdvectableScalarData(const scalarGrid3Ptr& scalarGrid3Data, FloatType initialVal = 0.0);
     size_t addAdvectableVectorGrid3Data(const vectorGrid3Ptr vectorGrid3Data, const vector3& initValue = vector3());
 
     const faceCenteredGrid3Ptr& velocity() const;
